// src/types.ts
export type UserDetails = {
  avatar_url: string;
  bio: String | undefined;
  blog: String | undefined;
  company: null;
  created_at: String | undefined;
  email: null;
  events_url: String | undefined;
  followers: any;
  followers_url: String | undefined;
  following: Number | undefined;
  following_url: String | undefined;
  gists_url: String | undefined;
  gravatar_id: String | undefined;
  hireable: null;
  html_url: String | undefined;
  id: Number;
  location: String | null;
  login: String | undefined;
  name: String | undefined;
  node_id: String | undefined;
  organizations_url: String | undefined;
  public_gists: Number | undefined;
  public_repos: any | undefined;
  received_events_url: String | undefined;
  repos_url: String | undefined;
  score: Number | undefined;
  site_admin: Boolean | undefined;
  starred_url: String | undefined;
  subscriptions_url: String | undefined;
  twitter_username: null;
  type: String | undefined;
  updated_at: String | undefined;
  url: String | undefined;
};